test(cli): fold long-running operation filter cases into filter_test.go
